// src/app/feature-modules/main/sidebarconfigure/tabletriggers/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymTrigger",
    PATH = "/main/configure/table-triggers",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "trigger_id",
    SORT_ORDER = "asc",
    SAVE = "Save",
    UPDATE = "Update"
}